Make dockerfile generators abstract
* This allows us to add a fast build dockerfile
generator that we can sub out based on the fast
build flag.
