+++ pkg/dockerfile/standard_generator_test.go
@@ -96,7 +96,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -131,7 +131,7 @@ predict: predict.py:Predictor
 `))
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -179,7 +179,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -235,7 +235,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -291,7 +291,7 @@ build:
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -383,7 +383,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 
@@ -486,7 +486,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(false)
 	actual, err := gen.GenerateDockerfileWithoutSeparateWeights()
@@ -521,7 +521,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(true)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -558,7 +558,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(true)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -609,7 +609,7 @@ predict: predict.py:Predictor
 		require.NoError(t, err)
 		require.NoError(t, conf.ValidateAndComplete(""))
 
-		gen, err := NewGenerator(conf, tmpDir)
+		gen, err := NewStandardGenerator(conf, tmpDir)
 		require.NoError(t, err)
 		gen.SetUseCogBaseImage(true)
 		_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -667,7 +667,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(true)
 	_, actual, _, err := gen.GenerateModelBaseWithSeparateWeights("r8.im/replicate/cog-test")
@@ -718,7 +718,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(true)
 	gen.SetStrip(true)
@@ -801,7 +801,7 @@ predict: predict.py:Predictor
 	require.NoError(t, err)
 	require.NoError(t, conf.ValidateAndComplete(""))
 
-	gen, err := NewGenerator(conf, tmpDir)
+	gen, err := NewStandardGenerator(conf, tmpDir)
 	require.NoError(t, err)
 	gen.SetUseCogBaseImage(true)
 	gen.SetStrip(true)
